E-commerceWebsite
A simple e-commerce website showcasing various necklaces using HTML, CSS, and JavaScript with Bootstrap. Users can view products, add them to a shopping cart, and see the total items and price. The design features a cherry red theme for the header and footer, with a responsive layout and stylish buttons. The shopping cart dynamically updates based on user interactions.